+++ src/components/week1/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Week 1 Components Index
+ * 
+ * This barrel file exports all Week 1 related components for clean imports.
+ * 
+ * Usage:
+ * ```ts
+ * import { Part1DataGovernance, Part2TechnologyMath } from '@/components/week1';
+ * ```
+ * 
+ * @module week1
+ */
+
+export { default as Part1DataGovernance } from './Part1DataGovernance';
+export { default as BusStopExample } from './BusStopExample';
+export { default as FluShotExample } from './FluShotExample';
+
+// Part2TechnologyMath will be added when extracted from Spring2026Week1.tsx